Extract mouse UI simulation validation
Move play-mode, request-option, and active-drag checks into a dedicated
validator while keeping public pause text on the use-case facade. Pass that
text explicitly to preserve dependency direction and every wire-visible
failure response.
