+++ Packages/src/Editor/FirstPartyTools/SimulateMouseUi/MouseUiSimulationValidator.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d72a04fed9944f7b9ad6228ef9b6b120
+guid: 10c5768ed7de48a1af6d0d6d3fb8132c
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
